Cleanup & bugfix at image_demo
