+++ demos/terminal_demo.py
@@ -7,7 +7,7 @@ from base.rotator import Rotator
 import pathfinders.dijkstra as Dijkstra
 import pathfinders.longest_path as LongestPath
 
-from demos.demo_utils_v2 import ALGORITHM_NAMES, str2bool, available_algorithm, available_exporter
+from demos.demo_utils import ALGORITHM_NAMES, str2bool, get_algorithm, get_exporter
 
 
 DEFAULT_EXPORTER = "UnicodeExporter"
@@ -28,8 +28,8 @@ if __name__ == "__main__":
 
     rows = args.rows
     columns = args.columns
-    algorithm = available_algorithm(args.algorithm, AVAILABLE_ALGORITHMS)
-    exporter = available_exporter(args.exporter, AVAILABLE_EXPORTERS)
+    algorithm = get_algorithm(args.algorithm, AVAILABLE_ALGORITHMS)
+    exporter = get_exporter(args.exporter, AVAILABLE_EXPORTERS)
     rotations = args.rotations
     pathfinding = args.pathfinding
     print("Algorithm: {}\nRows: {}\ncolumns: {}\nExporter: {}".format(args.algorithm, rows, columns, args.exporter))
